Refactor Check signature to middleware
